refactor(HitStrategy): 플레이어가 String이 아닌 Supplier<String>을 받도록 변경
re
